# riskcast-v16-main/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: RISKCAST V3 API
  description: |
    Enterprise Insurance-Grade Risk Intelligence Platform API
    
    ## Authentication
    All endpoints require JWT authentication via Bearer token in the Authorization header.
    
    ## Versioning
    API is versioned via URL path (e.g., `/api/v3/...`)
    
    ## Rate Limiting
    - Standard: 100 requests/minute
    - Authenticated: 1000 requests/minute
    
    ## Audit Trail
    All write operations are recorded in an immutable audit ledger with hash chain verification.
    
    ## Multi-Tenancy
    All operations are automatically scoped to the authenticated user's tenant.
    
    ## Permissions
    Endpoints require specific RBAC permissions. See individual endpoint documentation.
  version: 3.0.0
  contact:
    name: API Support
    email: api-support@riskcast.com
  license:
    name: Proprietary
    url: https://riskcast.com/terms
servers:
  - url: https://api.riskcast.com/api/v3
    description: Production
  - url: https://staging-api.riskcast.com/api/v3
    description: Staging
  - url: http://localhost:8000/api/v3
    description: Local Development
...
paths:
...
  /risk-assessments:
    post:
      tags: [Risk Assessment]
      summary: Create risk assessment
      description: |
        Create a new risk assessment for a shipment.
        
        The assessment is deterministic - same input will always produce same output.
        Results include risk score, VaR calculations, and risk factors breakdown.
        
        **Permission Required**: `RISK_WRITE`
      operationId: createRiskAssessment
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskAssessmentCreate'
            example:
              shipment:
                cargo_type: ELECTRONICS
                cargo_value_usd: 500000
                container_count: 10
                packaging_quality: EXCELLENT
              route:
                origin_port: CNSHA
                destination_port: NLRTM
                carrier_code: MAEU
                estimated_departure: "2024-02-01T10:00:00Z"
                estimated_arrival: "2024-03-01T10:00:00Z"
              coverage:
                coverage_type: ALL_RISK
                insured_value_cents: 50000000
                currency: USD
      responses:
        '201':
          description: Assessment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskAssessmentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
    
    get:
      tags: [Risk Assessment]
      summary: List risk assessments
      description: |
        List risk assessments with pagination and filtering.
        
        **Permission Required**: `RISK_READ`
      operationId: listRiskAssessments
      security:
        - BearerAuth: []
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
            minimum: 1
        - name: status
          in: query
          schema:
            type: string
            enum: [READY, PENDING, RUNNING, COMPLETED, FAILED]
      responses:
        '200':
          description: List of assessments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskAssessmentListResponse'
  /risk-assessments/{assessment_id}:
    get:
      tags: [Risk Assessment]
      summary: Get risk assessment
      description: |
        Get detailed information about a risk assessment.
        
        **Permission Required**: `RISK_READ`
      operationId: getRiskAssessment
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AssessmentId'
      responses:
        '200':
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskAssessmentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /risk-runs:
    get:
      tags: [Risk Runs]
      summary: List risk runs
      description: |
        List risk runs with filtering.
        
        **Permission Required**: `RISK_READ`
      operationId: listRiskRuns
      security:
        - BearerAuth: []
      parameters:
        - name: assessment_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELLED]
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
      responses:
        '200':
          description: List of risk runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskRunListResponse'
  /risk-runs/{run_id}:
    get:
      tags: [Risk Runs]
      summary: Get risk run
      description: |
        Get detailed information about a risk run including results.
        
        **Permission Required**: `RISK_READ`
      operationId: getRiskRun
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Risk run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RiskRunDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================================================
  # Underwriting
  # ===========================================================================
...
components:
...
  schemas:
    RiskAssessmentCreate:
      type: object
      required: [shipment, route, coverage]
      properties:
        shipment:
          type: object
          required: [cargo_type, cargo_value_usd]
          properties:
            cargo_type:
              type: string
              enum: [ELECTRONICS, MACHINERY, TEXTILES, FOOD_PERISHABLE, FOOD_DRY, CHEMICALS, GENERAL]
            cargo_value_usd:
              type: number
              minimum: 1000
            container_count:
              type: integer
              minimum: 1
            packaging_quality:
              type: string
              enum: [POOR, STANDARD, GOOD, EXCELLENT]
        route:
          type: object
          required: [origin_port, destination_port]
          properties:
            origin_port:
              type: string
              description: UN/LOCODE
              example: CNSHA
            destination_port:
              type: string
              description: UN/LOCODE
              example: NLRTM
            carrier_code:
              type: string
              example: MAEU
            estimated_departure:
              type: string
              format: date-time
            estimated_arrival:
              type: string
              format: date-time
        coverage:
          type: object
          required: [coverage_type, insured_value_cents]
          properties:
            coverage_type:
              type: string
              enum: [ALL_RISK, NAMED_PERILS, TOTAL_LOSS]
            insured_value_cents:
              type: integer
              minimum: 100000
            currency:
              type: string
              default: USD
    RiskAssessmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        created_by_user_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [READY, PENDING, RUNNING, COMPLETED, FAILED]
        input_schema_version:
          type: string
        input_hash:
          type: string
          description: SHA256 hash of canonical input
        shipment_id:
          type: string
          format: uuid
        corridor_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    RiskAssessmentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RiskAssessmentResponse'
        total:
          type: integer
        skip:
          type: integer
        limit:
          type: integer
...
    RiskRunResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        assessment_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [QUEUED, RUNNING, COMPLETED, FAILED, CANCELLED]
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    RiskRunDetailResponse:
      allOf:
        - $ref: '#/components/schemas/RiskRunResponse'
        - type: object
          properties:
            result:
              $ref: '#/components/schemas/RiskResult'
            provenance:
              $ref: '#/components/schemas/Provenance'
    RiskRunListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RiskRunResponse'
        total:
          type: integer
    RiskResult:
      type: object
      properties:
        risk_score:
          type: number
          minimum: 0
          maximum: 1
        var_95:
          type: number
          description: Value at Risk (95%)
        var_99:
          type: number
          description: Value at Risk (99%)
        expected_loss:
          type: number
        expected_loss_rate:
          type: number
        risk_factors:
          type: object
          additionalProperties:
            type: number
    Provenance:
      type: object
      description: Cryptographic provenance for auditability
      properties:
        model_version_id:
          type: string
          format: uuid
        model_version_name:
          type: string
        model_hash:
          type: string
        risk_run_id:
          type: string
          format: uuid
        input_hash:
          type: string
        result_hash:
          type: string
        seed:
          type: integer
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
        code:
          type: string

  # ===========================================================================
  # Parameters
  # ===========================================================================
  parameters:
    AssessmentId:
      name: assessment_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Risk assessment ID
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Risk run ID
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string

  # ===========================================================================
  # Security
  # ===========================================================================
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from authentication endpoint.
        Include in Authorization header as: `Bearer <token>`
security:
  - BearerAuth: []